+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Added a mapstructure DecodeHook for zap and zapcore types used in configuration
 
 ## [v0.1.1]
 - Rename Options to Config and use embedding to tidy up